feat: fix second page
