+++ shelter/package.json
@@ -7,12 +7,21 @@
     "test": "echo \"Error: no test specified\" && exit 1",
     "style": "sass --watch src/sass/style.scss style.css",
     "dev": "webpack --mode development",
-    "build": "webpack --mode production"
+    "build": "webpack --mode production",
+    "webwatch": "webpack --watch"
   },
   "author": "mariia",
   "license": "ISC",
   "dependencies": {
     "webpack": "^5.95.0",
     "webpack-cli": "^5.1.4"
+  },
+  "devDependencies": {
+    "css-loader": "^7.1.2",
+    "file-loader": "^6.2.0",
+    "html-webpack-plugin": "^5.6.2",
+    "sass": "^1.80.3",
+    "sass-loader": "^16.0.2",
+    "style-loader": "^4.0.0"
   }
 }
