+++ shelter/pets.html
@@ -121,7 +121,6 @@
             </div>
         </div> -->
     </div>
-    <script src="src/js/burger-friends.js"></script>
-    <script src="src/js/pagination.js"></script>
+    <script type="module" src="src/pagination.js"></script>
 </body>
 </html>
