+++ shelter/webpack.config.js
@@ -1,0 +1,24 @@
+const path = require('path');
+
+module.export {
+    
+        entry: {'./src/index.js'},
+        output: {
+            filename: 'bundle.js',
+            path: path.resolve(__dirname, 'dist'),
+        
+
+    // module: {
+    //     rules: [
+    //         {
+    //             test: /\.css$/,
+    //             use: [
+    //                 'style-loader', 'css-loader', 'sass-loader'
+    //             ],
+    //         }
+    //     ]
+    // }    
+    }
+
+    return config;
+}
